requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
botocore==1.20.86
celery==5.1.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
clipboard==0.0.4
Django==3.2.3
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-redis==5.0.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
image==1.5.33
jmespath==0.10.0
kombu==5.1.0
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyperclip==1.8.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
